+++ anon-service.sh
@@ -2,10 +2,10 @@
 
 # #####################################################################
 # anon-service.sh
-# version 1.1
+# version 1.2
 # 
-# Transparent proxy through Tor with DNSCrypt and Anonymized DNS 
-# feature enabled.
+# Transparent proxy through Tor and optionally DNSCrypt with Anonymized 
+# DNS feature enabled.
 #
 # Copyright (C) 2020-2021 Bit4mind
 #
@@ -31,8 +31,8 @@ repo=/etc/apt/sources.list.d/tor.list
 ## DNSCrypt-proxy release
 dnscrel="2.0.45"
 ## If necessary, change the path according to your system
-netman=/etc/NetworkManager/NetworkManager.conf
-resolved=/etc/systemd/resolved.conf
+export netman=/etc/NetworkManager/NetworkManager.conf
+export resolved=/etc/systemd/resolved.conf
 tor=/etc/tor/torrc
 unbound=/etc/unbound/unbound.conf
 
@@ -52,7 +52,7 @@ echo "              ▒██████▒▒░▒████▒░██▓
 echo "                    ░           ░           ░       ░ by bit4mind  ";
 echo " ";
 echo "   0. Check dependencies and download upgraded services";
-echo "   1. Set servers/relays for anonymized DNS feature and configure other services";
+echo "   1. Choose transparent proxy type and configure related services";
 echo "   2. Start/Restart anon-service";
 echo "   3. Execute all tasks above";
 echo "   4. Close this window";
@@ -85,22 +85,22 @@ start_service
 menu
 ;;
 4)
-if [ -s "cpath" ]; then
+if [ -f "cpath" ]; then
 mv cpath $root/ > /dev/null 2>&1
 fi
 wmctrl -c :ACTIVE:
 ;;
 5)
-if [ ! -s "$root/dnscrypt-proxy.toml" ]; then
+if [ -f "cpath" ]; then
+mv cpath $root/ > /dev/null 2>&1
+fi
+if [ ! -f "$root/stp-service" ]; then
 echo "";
 echo "==> Sorry! Your system is not ready to start the service...";
 echo "==> Please, check if you have installed the necessary files";
 sleep 7
 menu
-exit
-fi
-if [ -s "cpath" ]; then
-mv cpath $root/ > /dev/null 2>&1
+return 1
 fi
 cd $root
 cp resolved.conf.temp $resolved > /dev/null 2>&1
@@ -131,8 +131,11 @@ echo "service dnscrypt-proxy stop > /dev/null 2>&1" >> /etc/network/if-up.d/anon
 echo "service unbound stop > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "killall unbound tor dnscrypt-proxy > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "chown -R $owner:$owner $root" >> /etc/network/if-up.d/anon-service
+selected_service=$(cat $root/stp-service)
+if (( $selected_service == 1 )); then
 echo "nohup su - $owner -c \"./dnscrypt-proxy\" > /dev/null 2>&1 &" >> /etc/network/if-up.d/anon-service
 echo "sleep 1s" >> /etc/network/if-up.d/anon-service
+fi
 echo "rm $root/notices.log > /dev/null 2>&1" >> /etc/network/if-up.d/anon-service
 echo "touch $root/notices.log" >> /etc/network/if-up.d/anon-service
 echo "chown anon-service:anon-service $root/notices.log" >> /etc/network/if-up.d/anon-service
@@ -156,7 +159,11 @@ echo "iptables -t nat -A OUTPUT -d \$_virt_addr -p tcp -m tcp --tcp-flags FIN,SY
 echo "iptables -A OUTPUT -m state --state INVALID -j DROP" >> /etc/network/if-up.d/anon-service
 echo "iptables -A OUTPUT -m state --state ESTABLISHED,RELATED -j ACCEPT" >> /etc/network/if-up.d/anon-service
 echo "iptables -t nat -A OUTPUT -m owner --uid-owner \$_user_uid -j RETURN" >> /etc/network/if-up.d/anon-service
+if (( $selected_service == 0 )); then
 echo "iptables -t nat -A OUTPUT -p udp --dport 53 -j REDIRECT --to-ports 53" >> /etc/network/if-up.d/anon-service
+else
+echo "iptables -t nat -A OUTPUT -p udp --dport 53 -j REDIRECT --to-ports 5353" >> /etc/network/if-up.d/anon-service
+fi
 echo "for _clearnet in \$_non_tor; do" >> /etc/network/if-up.d/anon-service
 echo "iptables -t nat -A OUTPUT -d \$_clearnet -j RETURN" >> /etc/network/if-up.d/anon-service
 echo "done" >> /etc/network/if-up.d/anon-service
@@ -181,7 +188,9 @@ echo "sleep 1s" >> /etc/network/if-up.d/anon-service
 echo "iptables -P FORWARD DROP" >> /etc/network/if-up.d/anon-service
 echo "iptables -P INPUT DROP" >> /etc/network/if-up.d/anon-service
 echo "iptables -P OUTPUT DROP" >> /etc/network/if-up.d/anon-service
+if (( $selected_service == 1 )); then
 echo "unbound" >> /etc/network/if-up.d/anon-service
+fi
 echo "echo \"+++ anon-service started +++\"" >> /etc/network/if-up.d/anon-service
 chown root:root /etc/network/if-up.d/anon-service
 chmod 755 /etc/network/if-up.d/anon-service
@@ -189,10 +198,11 @@ chmod +x /etc/network/if-up.d/anon-service
 echo "";
 echo "==> Now you are ready to go! If you haven't set 127.0.0.1 in your DNS"; 
 echo "==> setting, do it and restart your connection or reboot your system.";
-exit 0
+echo "";
+exit
 ;;
 6)
-if [ -s "cpath" ]; then
+if [ -f "cpath" ]; then
 mv cpath $root/ > /dev/null 2>&1
 fi
 shutdown_service
@@ -202,14 +212,15 @@ menu
 cleanall
 ;;
 8)
-if [ -s "cpath" ]; then
+if [ -f "cpath" ]; then
 mv cpath $root/ > /dev/null 2>&1
 fi
-service dnscrypt-proxy stop > /dev/null 2>&1
-sleep 3
-if ! pgrep -x "dnscrypt-proxy" > /dev/null; then
+#service dnscrypt-proxy stop > /dev/null 2>&1
+#sleep 3
+if ! pgrep -x "tor" > /dev/null; then
 echo "";
 echo "==> Service is not running!"
+sleep 3
 menu
 else
 echo "";
@@ -219,7 +230,7 @@ menu
 fi
 ;;
 9)
-if [ -s "cpath" ]; then
+if [ -f "cpath" ]; then
 mv cpath $root/ > /dev/null 2>&1
 fi
 touch /usr/bin/anon-service > /dev/null 2>&1
@@ -246,7 +257,7 @@ echo "";
 echo "==> Please, firstly remove all files and settings via dedicated option";
 sleep 7
 menu
-exit
+return 1
 fi
 clear
 echo "==> Checking dependencies and preparing the system"
@@ -385,20 +396,37 @@ cd $(cat $root/cpath)
 ## CONFIGURING SERVICES
 ##
 configure(){
+if [ -f "cpath" ]; then
+mv cpath $root/ > /dev/null 2>&1
+fi
 if [ ! -s "$root/dnscrypt-proxy.toml.bak" ]; then
 echo "";
 echo "==> Sorry! Your system is not ready to complete this action";
 echo "==> Please, check if you have installed the necessary files";
 sleep 7
 menu
-exit
-fi
-if [ -s "cpath" ]; then
-mv cpath $root/ > /dev/null 2>&1
+return 1
 fi
 ### Disable tor and unbound starting at boot time
 systemctl disable unbound > /dev/null 2>&1
 systemctl disable tor > /dev/null 2>&1
+clear
+echo "==> Which type of transparent proxy do you prefer to use?";
+echo "      1. Standard transparent proxy";
+echo "      2. Trasparent proxy with DNSCrypt and Anonymized DNS feature";
+echo " "
+echo -n  " Choose: ";
+read -e choose
+case "$choose" in 
+1)
+rm $root/stp-service > /dev/null 2>&1
+touch $root/stp-service
+echo "1" > $root/stp-service
+;;
+2)
+rm $root/stp-service > /dev/null 2>&1
+touch $root/stp-service
+echo "0" > $root/stp-service 
 ### Configuring dnscrypt_proxy
 rm $root/dnscrypt-proxy.toml > /dev/null 2>&1
 cp $root/dnscrypt-proxy.toml.bak $root/dnscrypt-proxy.toml
@@ -409,44 +437,80 @@ clear
 echo "==> Please enter the name of the first resolver to use, only ipv4!";
 echo -n "    First server: ";
 read -e server1
-echo " ";
+echo "";
+if ! grep "\<$server1\>" $root/public-resolvers.md > /dev/null; then
+echo "==> Server not found! Please retry";
+killall gedit > /dev/null 2>&1
+sleep 3
+configure
+return 1
+fi
 echo "==> Please enter the name of the second resolver to use, only ipv4!";
 echo -n "    Second server: ";
 read -e server2
+if ! grep "\<$server2\>" $root/public-resolvers.md > /dev/null; then
+echo "==> Server not found! Please retry";
+killall gedit > /dev/null 2>&1
+sleep 3
+configure
+return 1
+fi
 echo "==> Opening file contain relays";
 killall gedit > /dev/null 2>&1
 xterm -T "Relay" -e "gedit $root/relays.md" &
 clear
 echo "==> Carefully choose relays/servers so that they are run by different entities!";
-echo " ";
+echo "";
 echo "==> Please enter the name of the first realy to use!";
 echo -n "    First relay for the first server: ";
 read -e relay1
-echo " ";
+echo "";
+if ! grep "\<$relay1\>" $root/relays.md > /dev/null; then
+echo "==> Relay not found! Please retry";
+killall gedit > /dev/null 2>&1
+sleep 3
+configure
+return 1
+fi
 echo "==> Please enter the name of the second relay to use!";
 echo -n "    Second relay for the first server: ";
 read -e relay2
-echo " ";
+echo "";
+if ! grep "\<$relay2\>" $root/relays.md > /dev/null; then
+echo "==> Relay not found! Please retry";
+killall gedit > /dev/null 2>&1
+sleep 3
+configure
+return 1
+fi
 echo "==> Please enter the name of the third resolver to use!";
 echo -n "    First relay for the second server: ";
 read -e relay3
-echo " ";
+echo "";
+if ! grep "\<$relay3\>" $root/relays.md > /dev/null; then
+echo "==> Relay not found! Please retry";
+killall gedit > /dev/null 2>&1
+sleep 3
+configure
+return 1
+fi
 echo "==> Please enter the name of the fourth resolver to use!";
 echo -n "    Second relay for the second server: ";
 read -e relay4
+echo "";
+if ! grep "\<$relay4\>" $root/relays.md > /dev/null; then
+echo "==> Relay not found! Please retry";
+killall gedit > /dev/null 2>&1
+sleep 3
+configure
+return 1
+fi
 killall gedit > /dev/null 2>&1
 clear
 echo "==> Configuring other services";
 sed -i "1iforce_tcp = true" $root/dnscrypt-proxy.toml
 sed -i "2iserver_names = ['$server1', '$server2']" $root/dnscrypt-proxy.toml
 sed -i "s/127.0.0.1:53/127.0.0.1:10000/g; s/9.9.9.9/208.67.222.222/g; s/8.8.8.8/208.67.220.220/g; s/require_dnssec = false/require_dnssec = true/g; s/force_tcp = false/#force_tcp = false/g; s/\[anonymized_dns\]/\[anonymized_dns\]\nroutes = \[\n{ server_name='$server1', via=\[\'$relay1\', \'$relay2\'\] },\n{ server_name=\'$server2\', via=[\'$relay3\', \'$relay4\'] }\n\]/g; s/skip_incompatible = false/skip_incompatible = true/g" $root/dnscrypt-proxy.toml
-### Configuring Tor
-cp $tor $root/torrc
-echo "Log notice file $root/notices.log" >> $root/torrc
-echo "VirtualAddrNetworkIPv4 10.192.0.0/10" >> $root/torrc
-echo "AutomapHostsOnResolve 1" >> $root/torrc
-echo "TransPort 9040 IsolateClientAddr IsolateClientProtocol IsolateDestAddr IsolateDestPort" >> $root/torrc
-echo "DNSPort 5353" >> $root/torrc
 ### Configuring unbound
 unbound-anchor > /dev/null 2>&1
 sleep 1
@@ -463,7 +527,22 @@ echo "    forward-addr: 127.0.0.1@5353" >> $unbound
 echo "forward-zone:" >> $unbound
 echo "   name: \".\"" >> $unbound
 echo "   forward-addr: 127.0.0.1@10000" >> $unbound
+;;
+*)
+echo "";
+echo "==> Are you serious?"
+sleep 5
+configure
+esac
+### Configuring Tor
+cp $tor $root/torrc
+echo "Log notice file $root/notices.log" >> $root/torrc
+echo "VirtualAddrNetworkIPv4 10.192.0.0/10" >> $root/torrc
+echo "AutomapHostsOnResolve 1" >> $root/torrc
+echo "TransPort 9040 IsolateClientAddr IsolateClientProtocol IsolateDestAddr IsolateDestPort" >> $root/torrc
+echo "DNSPort 5353" >> $root/torrc
 ### Disabling dnsmasq and configure Network-Manager and systemd-resolved
+rm $root/NetworkManager.conf.temp > /dev/null 2>&1
 cp $netman.bak $root/NetworkManager.conf.temp
 cd $root
 chown $USER:$USER NetworkManager.conf.temp
@@ -475,6 +554,55 @@ chown $USER:$USER resolved.conf.temp
 sed -i 's/^DNSStubListener=yes/#&/' resolved.conf.temp
 echo "DNSStubListener=no" >> resolved.conf.temp
 fi
+rm $root/iptables_rules.sh > /dev/null 2>&1
+touch $root/iptables_rules.sh
+### Configuring basic iptables rules
+### Reference: https://trac.torproject.org/projects/tor/wiki/doc/TransparentProxy
+echo "#!/bin/bash" > $root/iptables_rules.sh
+# Destinations you don't want routed through Tor
+echo "_non_tor=\"127.0.0.0/8 10.0.0.0/8 172.16.0.0/12 192.168.0.0/16\"" >> $root/iptables_rules.sh
+# The UID that Tor runs as (varies from system to system)
+echo "_user_uid=\"999\"" >> $root/iptables_rules.sh
+# Tor's VirtualAddrNetworkIPv4
+echo "_virt_addr=\"10.192.0.0/10\"" >> $root/iptables_rules.sh
+# Tor's TransPort
+echo "_trans_port=\"9040\"" >> $root/iptables_rules.sh
+echo "iptables -F" >> $root/iptables_rules.sh
+echo "iptables -t nat -F" >> $root/iptables_rules.sh
+echo "iptables -t nat -A OUTPUT -d \$_virt_addr -p tcp -m tcp --tcp-flags FIN,SYN,RST,ACK SYN -j REDIRECT --to-ports \$_trans_port" >> $root/iptables_rules.sh
+echo "iptables -A OUTPUT -m state --state INVALID -j DROP" >> $root/iptables_rules.sh
+echo "iptables -A OUTPUT -m state --state ESTABLISHED,RELATED -j ACCEPT" >> $root/iptables_rules.sh
+echo "iptables -t nat -A OUTPUT -m owner --uid-owner \$_user_uid -j RETURN" >> $root/iptables_rules.sh
+if ( grep "1" $root/stp-service ); then
+echo "iptables -t nat -A OUTPUT -p udp --dport 53 -j REDIRECT --to-ports 5353" >> $root/iptables_rules.sh
+else
+echo "iptables -t nat -A OUTPUT -p udp --dport 53 -j REDIRECT --to-ports 53" >> $root/iptables_rules.sh
+fi
+echo "for _clearnet in \$_non_tor; do" >> $root/iptables_rules.sh
+echo "iptables -t nat -A OUTPUT -d \$_clearnet -j RETURN" >> $root/iptables_rules.sh
+echo "done" >> $root/iptables_rules.sh
+echo "sleep 5" >> $root/iptables_rules.sh
+echo "iptables -A INPUT -m state --state ESTABLISHED -j ACCEPT" >> $root/iptables_rules.sh
+echo "iptables -A INPUT -i lo -j ACCEPT" >> $root/iptables_rules.sh
+echo "for _lan in \$_non_tor; do" >> $root/iptables_rules.sh
+echo "iptables -A INPUT -s \$_lan -j ACCEPT" >> $root/iptables_rules.sh
+echo "done" >> $root/iptables_rules.sh
+echo "sleep 5" >> $root/iptables_rules.sh
+echo "iptables -A INPUT -j DROP" >> $root/iptables_rules.sh
+echo "iptables -A FORWARD -j DROP" >> $root/iptables_rules.sh
+echo "for _clearnet in \$_non_tor; do" >> $root/iptables_rules.sh
+echo "iptables -A OUTPUT -d \$_clearnet -j ACCEPT" >> $root/iptables_rules.sh
+echo "done" >> $root/iptables_rules.sh
+echo "sleep 3" >> $root/iptables_rules.sh
+echo "iptables -t nat -A OUTPUT -p tcp --syn -j REDIRECT --to-ports \$_trans_port" >> $root/iptables_rules.sh
+echo "iptables -A OUTPUT -m owner --uid-owner \$_user_uid -j ACCEPT" >> $root/iptables_rules.sh
+echo "sleep 2" >> $root/iptables_rules.sh
+echo "iptables -A OUTPUT -j DROP" >> $root/iptables_rules.sh
+echo "sleep 1" >> $root/iptables_rules.sh
+echo "iptables -P FORWARD DROP" >> $root/iptables_rules.sh
+echo "iptables -P INPUT DROP" >> $root/iptables_rules.sh
+echo "iptables -P OUTPUT DROP" >> $root/iptables_rules.sh
+chmod +x $root/iptables_rules.sh
 cd $(cat $root/cpath)
 }
 ##
@@ -482,15 +610,15 @@ cd $(cat $root/cpath)
 ##
 start_service(){
 ### Checking for required files
-if [ ! -s "$root/dnscrypt-proxy.toml" ]; then
+if [ ! -s "$root/stp-service" ]; then
 echo "";
 echo "==> Sorry! Your system is not ready to start the service...";
 echo "==> Please, check if you have installed the necessary files";
 sleep 7
 menu
-exit
+return 1
 fi
-if [ -s "cpath" ]; then
+if [ -f "cpath" ]; then
 mv cpath $root/ > /dev/null 2>&1
 fi
 ### Firewall flush
@@ -514,7 +642,6 @@ service bind stop > /dev/null 2>&1
 service dnscrypt-proxy stop > /dev/null 2>&1
 killall dnsmasq bind > /dev/null 2>&1
 sleep 1
-cd $root
 service tor stop > /dev/null 2>&1
 service dnscrypt-proxy stop > /dev/null 2>&1
 service unbound stop > /dev/null 2>&1
@@ -531,6 +658,30 @@ service systemd-resolved restart
 service network-manager restart
 sleep 10
 chown -R $owner:$owner $root
+## Restore original files automatically at shutdown
+if ! pgrep -f "restoring_orig.sh"  > /dev/null; then
+rm restoring_orig.sh > /dev/null 2>&1 
+touch restoring_orig.sh
+echo "#!/bin/bash" > restoring_orig.sh
+echo "restoring_script() {" >> restoring_orig.sh 
+echo "if [ ! -f /etc/network/if-up.d/anon-service ]; then" >> restoring_orig.sh
+echo "cp $root/resolved.bak $resolved" >> restoring_orig.sh
+echo "cp $netman.bak $netman" >> restoring_orig.sh
+echo "fi" >> restoring_orig.sh
+echo "exit" >> restoring_orig.sh
+echo "}" >> restoring_orig.sh
+echo "while :" >> restoring_orig.sh
+echo "do" >> restoring_orig.sh
+echo "trap restoring_script SIGINT SIGTERM" >> restoring_orig.sh
+echo "sleep 1" >> restoring_orig.sh
+echo "done" >> restoring_orig.sh
+chmod +x restoring_orig.sh
+xterm -e nohup ./restoring_orig.sh
+fi
+## Start selected transparent proxy
+active_service=$(cat $root/stp-service)
+case $active_service in
+"0")
 nohup su - $owner -c "./dnscrypt-proxy" > /dev/null 2>&1 &
 sleep 1
 rm $root/notices.log > /dev/null 2>&1
@@ -548,47 +699,8 @@ fi
 sleep 1
 done
 rm $root/notices.log > /dev/null 2>&1
-### Configuring basic iptables rules
-### Reference: https://trac.torproject.org/projects/tor/wiki/doc/TransparentProxy
-# Destinations you don't want routed through Tor
-_non_tor="127.0.0.0/8 10.0.0.0/8 172.16.0.0/12 192.168.0.0/16"
-# The UID that Tor runs as (varies from system to system)
-_user_uid="999"
-# Tor's VirtualAddrNetworkIPv4
-_virt_addr="10.192.0.0/10"
-# Tor's TransPort
-_trans_port="9040"
-iptables -F
-iptables -t nat -F
-iptables -t nat -A OUTPUT -d $_virt_addr -p tcp -m tcp --tcp-flags FIN,SYN,RST,ACK SYN -j REDIRECT --to-ports $_trans_port
-iptables -A OUTPUT -m state --state INVALID -j DROP
-iptables -A OUTPUT -m state --state ESTABLISHED,RELATED -j ACCEPT
-iptables -t nat -A OUTPUT -m owner --uid-owner $_user_uid -j RETURN
-iptables -t nat -A OUTPUT -p udp --dport 53 -j REDIRECT --to-ports 53
-for _clearnet in $_non_tor; do
-iptables -t nat -A OUTPUT -d $_clearnet -j RETURN
-done
-sleep 5
-iptables -A INPUT -m state --state ESTABLISHED -j ACCEPT
-iptables -A INPUT -i lo -j ACCEPT
-for _lan in $_non_tor; do
-iptables -A INPUT -s $_lan -j ACCEPT
-done
-sleep 5
-iptables -A INPUT -j DROP
-iptables -A FORWARD -j DROP
-for _clearnet in $_non_tor; do
-iptables -A OUTPUT -d $_clearnet -j ACCEPT
-done
-sleep 3
-iptables -t nat -A OUTPUT -p tcp --syn -j REDIRECT --to-ports $_trans_port
-iptables -A OUTPUT -m owner --uid-owner $_user_uid -j ACCEPT
-sleep 2
-iptables -A OUTPUT -j DROP
-sleep 1
-iptables -P FORWARD DROP
-iptables -P INPUT DROP
-iptables -P OUTPUT DROP
+cd $root
+./iptables_rules.sh
 unbound &
 ### Checking services
 if ! pgrep -x "tor" > /dev/null; then
@@ -596,22 +708,56 @@ echo "==> Sorry! No connection to TOR...Please, report this issue to the project
 sleep 7
 shutdown_service
 menu
+return 1
 fi
 if ! pgrep -x "dnscrypt-proxy" > /dev/null; then
 echo "==> Sorry! Dnscrypt-proxy isn't running...Please, report this issue to the project";
 sleep 7
 shutdown_service
 menu
+return 1
 fi
 if ! pgrep -x "unbound" > /dev/null; then
 echo "==> Sorry! Unbound isn't running...Please, report this issue to the project";
 sleep 7
 shutdown_service
 menu
+return 1
 else
 echo "==> Congratulations! Your system is configurated to use Tor and DNSCrypt";
 sleep 5
 fi
+;;
+"1")
+rm $root/notices.log > /dev/null 2>&1
+touch $root/notices.log
+chown anon-service:anon-service $root/notices.log
+nohup su - $owner -c "tor -f $root/torrc" > /dev/null 2>&1 &
+echo "==> Checking connection to Tor";
+SECONDS=0
+secs=30
+while (( SECONDS < secs ));
+do
+if (grep -Fq "100%" $root/notices.log ); then 
+break
+fi
+sleep 1
+done
+rm $root/notices.log > /dev/null 2>&1
+cd $root
+./iptables_rules.sh
+### Checking services
+if ! pgrep -x "tor" > /dev/null; then
+echo "==> Sorry! No connection to TOR...Please, report this issue to the project";
+sleep 7
+shutdown_service
+menu
+return 1
+else
+echo "==> Congratulations! Your system is configurated to use Tor";
+sleep 5
+fi
+esac
 cd $(cat $root/cpath)
 }
 ##
@@ -621,7 +767,7 @@ shutdown_service(){
 clear
 service dnscrypt-proxy stop > /dev/null 2>&1
 sleep 3
-if ! pgrep -x "dnscrypt-proxy" > /dev/null; then
+if ! pgrep -x "tor" > /dev/null; then
 echo "==> Restoring original files"; 
 sleep 7
 else
@@ -632,7 +778,7 @@ rm $root/tor.txt > /dev/null 2>&1
 service dnscrypt-proxy stop > /dev/null 2>&1
 service tor stop > /dev/null 2>&1
 service unbound stop > /dev/null 2>&1
-killall unbound tor dnscrypt-proxy > /dev/null 2>&1
+killall unbound tor dnscrypt-proxy restoring_orig.sh > /dev/null 2>&1
 cp $root/resolved.bak $resolved > /dev/null 2>&1
 cp $netman.bak $netman > /dev/null 2>&1
 if [ -s "/etc/network/if-up.d/anon-service" ]; then
@@ -666,7 +812,7 @@ echo "==> Removing anon-service files and settings from system";
 service tor stop > /dev/null 2>&1
 service dnscrypt-proxy stop > /dev/null 2>&1
 service unbound stop > /dev/null 2>&1
-killall unbound tor dnscrypt-proxy > /dev/null 2>&1
+killall unbound tor dnscrypt-proxy restoring_orig.sh > /dev/null 2>&1
 if [[ -s "$root/resolved.bak" ]]; then
 cp $root/resolved.bak $resolved > /dev/null 2>&1
 service systemd-resolved restart