+++ CHANGELOG.md
@@ -2,6 +2,21 @@ CHANGELOG
 
 All notable changes to this project will be documented in this file.
 
+UnReleased
+1.2 - 2021-04-19
+
+Added
+
+    Transparent proxy without DNSCrypt option
+    Restore original files automatically at shutdown
+    More input controls
+    
+
+Changed
+
+    Second option (1) in main menu
+    
+
 Released
 1.1 - 2021-01-02